requirements.txt
fastapi
uvicorn
torch
pillow
requests
firebase-admin
cloudinary
ultralytics